+++ app/database/migrations/2024_03_02_012358_add_spymaster_to_realms_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddSpymasterToRealmsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('realms', function (Blueprint $table) {
+            $table->unsignedInteger('spymaster_dominion_id')->nullable()->after('general_dominion_id');
+            $table->foreign('spymaster_dominion_id')->references('id')->on('dominions');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('realms', function (Blueprint $table) {
+            $table->dropForeign('realms_spymaster_dominion_id_foreign');
+            $table->dropColumn('spymaster_dominion_id');
+        });
+    }
+}