+++ src/Services/Dominion/Actions/InvadeActionService.php
@@ -493,6 +493,11 @@ class InvadeActionService
                 $dominion->prestige += $attackerPrestigeChange;
 
             } else {
+                // Royal Court: +2
+                if ($target->user_id !== null && $dominion->getCourtSeat() == 'general') {
+                    $attackerPrestigeChange += 2;
+                }
+
                 // todo: possible bug if all 12hr units die (somehow) and only 9hr units survive, prestige gets returned after 12 hrs, since $units is input, not surviving units. fix?
                 $slowestTroopsReturnHours = $this->invasionService->getSlowestUnitReturnHours($dominion, $units);
                 $this->queueService->queueResources(