+++ src/Services/Realm/HistoryService.php
@@ -12,6 +12,7 @@ class HistoryService
     public const EVENT_ACTION_DECLARE_WAR = 'declare war';
     public const EVENT_ACTION_CANCEL_WAR = 'cancel war';
     public const EVENT_ACTION_APPOINT_GENERAL = 'appointed general';
+    public const EVENT_ACTION_APPOINT_SPYMASTER = 'appointed spymaster';
     public const EVENT_ACTION_APPOINT_MAGISTER = 'appointed magister';
     public const EVENT_ACTION_APPOINT_MAGE = 'appointed mage';
     public const EVENT_ACTION_APPOINT_JESTER = 'appointed jester';
@@ -97,6 +98,7 @@ class HistoryService
                 'round_id',
                 'monarch_dominion_id',
                 'general_dominion_id',
+                'spymaster_dominion_id',
                 'magister_dominion_id',
                 'mage_dominion_id',
                 'jester_dominion_id',