+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes relevant to players in this project will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Added
+- Spymaster
+  - New role in the Royal Court
+  - Can post recurring bounties
+  - Can post black op bounties
+- Royal Court perks (not active during protection)
+  - Monarch: +2% food production
+  - General: +2 prestige gained against player targets (as opposed to bots)
+  - Spymaster: +5% spy success chance
+  - Grand Magister: -2.5% spell mana costs
+  - Court Mage: +0.2 wizard strength recovery per hour
+  - Jester: +1 morale recovery per hour
+
 ### Fixed
 - The blank space in the header no longer linkts to the homepage